better fix for the infinite loop
"latest" keyword now only activates queries if the current message is
the message being filtered for. Thus the latest query allows us to avoid
the infinite loop situation where a whenever condition would be
eroniously true. I would love a static check for things like this... (ie
a check for a non-latest query without an every or other query in it.)